Never try to update read-only, system-populated metadata fields.
If the user asks us to update them, we will silently ignore the
changes. The API server would ignore the changes anyway, but we
need to account for that in advance because otherwise we'll either
send needless no-op updates (for InPlace strategies) or needlessly
delete and recreate objects (for Recreate strategies).