refactor: #163 - Introduce controller runtime for watching top-level CRDs
Metacontroller, being an older project, contains a lot of boilerplate
for watching and fetching kubernetes resources. Most of that logic
has been extracted into sigs.k8s.io/controller-runtime.

This commit introduces controller-runtime for reconciling top level
resources, that is DecoratorControllers and CompositeControllers.
Metacontroller creates one controller for each of these resources when
it starts. The scope of this commit is refactoring these particular controllers
to take advantage of the machinery provided by controller-runtime.

In addition to the two static controllers, metacontroller creates a new dynamic
controller for each instance of DecoratorController and CompositeController.
These dynamic controllers have been left untouched and their refactoring
to controller-runtime, or to any other library for that manner, is out of scope
for this particular commit. That work would be done in a subsequent commit.

Since controller-runtime uses a dynamic client, these cnages make the
informes and listers for CompositeController and DecoratorController
under client/generated obsolete. Their removal can also be done as follow up
once this commit has been merged.

Signed-off-by: Filip Petkovski <filip.petkovsky@gmail.com>