perf(k8sClient): use a cache-based version of k8s client
The current metacontrollers use a plain controller-runtime client when talking to the API server.
As a result, each read and list action goes directly to the server instead of using a cache.

This commmit initializes the client to be the same as the one from the controller manager.
In this way, reading actions will always be backed by a cache, while write actions
will be executed directly against the API server.

Signed-off-by: fpetkovski <filip.petkovsky@gmail.com>