feat: make ownerReferences block controller field configurable
When decorator and composite controllers create child objects in
response to the user's sync hook, it appends an `ownerReferences` block
to the metadata of all children/attachments. At present that block has
the `controller` field set to `true`.

I have a use-case in which I don't want that set to `true` because I'm
creating objects that need to be owned and controlled by another
controller: cluster-api resources. With the current functionality
cluster-api will refuse to set its own ownership reference because there
is already one there with `controller: true` (which is correct
behaviour).

chore: update CRD manifests with new field
Signed-off-by: Luke Bond <luke.n.bond@gmail.com>